删除宿舍楼的时候，如果该宿舍楼内有学生，返回错误信息，不删除

+++ server/src/controller/building_controller.js
@@ -15,9 +15,15 @@ module.exports = {
   delBuilding: async function(id) {
     const { setStudentRoomNull } = require("./user_controller")
     const students = await this.getStudents(id)
-    students.forEach(student => {
-      setStudentRoomNull(student.id)
-    })
+    // students.forEach(student => {
+    //   setStudentRoomNull(student.id)
+    // })
+    if (students.length > 0) {
+      return {
+        code: 1,
+        msg: "该宿舍楼下有学生，请先将学生移出宿舍楼"
+      }
+    }
     return await Building.destroy({ where: { id } })
   }
 }
